refactor: validate 분리
- 기존의 vlidationUtils 에서 validations로 변경
- validations을 validateDomain, validateUtils, validateUI로 분리
