+++ src/Validation/validateUtils.js
@@ -1,45 +1,15 @@
-import { LOTTO_DEFINITION } from '../Domain/Constant/definition.js';
+import { hasEmptySpace, isInteger } from './validations.js';
 
-export const hasEmptySpace = (input) => {
-  return input.includes(' ') || input.trim() === '';
+import { COMMON_ERROR_MESSAGE } from '../View/Constant/errorMessage.js';
+
+export const validateEmptySpace = (input) => {
+  if (hasEmptySpace(input)) {
+    throw new Error(COMMON_ERROR_MESSAGE.NO_EMPTY_SPACE);
+  }
 };
 
-export const isInteger = (input) => {
-  return Number.isInteger(input);
-};
-
-export const isValidPurchaseAmountUnit = (input) => {
-  return input % LOTTO_DEFINITION.ONE_PRICE === 0;
-};
-
-export const isValidPurchaseAmountRange = (input) => {
-  return (
-    input >= LOTTO_DEFINITION.ONE_PRICE && input <= LOTTO_DEFINITION.MAX_PRICE
-  );
-};
-
-export const hasValidLength = (input) => {
-  return LOTTO_DEFINITION.NUMBER_COUNTS === new Set(input).size;
-};
-
-export const hasNoDuplicate = (input) => {
-  return input.length === new Set(input).size;
-};
-
-export const isInValidRange = (input) => {
-  return (
-    input >= LOTTO_DEFINITION.MIN_NUMBER && input <= LOTTO_DEFINITION.MAX_NUMBER
-  );
-};
-
-export const hasNoEmptySpaceInArray = (input) => {
-  return !input.some(hasEmptySpace);
-};
-
-export const hasNoDuplicateBonusNumber = (input, winningNumbersInput) => {
-  return !winningNumbersInput.includes(input);
-};
-
-export const isValidRetryInput = (input) => {
-  return input === 'y' || input === 'n';
+export const validateInteger = (input) => {
+  if (!isInteger(input)) {
+    throw new Error(COMMON_ERROR_MESSAGE.NOT_INTEGER);
+  }
 };
